+++ GameEngine/GameEngineFBXWindow.cpp
@@ -366,7 +366,6 @@ void GameEngineFBXWindow::OnGUI()
 		}
 	}
 
-
 	if (nullptr != SelectMesh
 		&& 0 != SelectMesh->GetMeshSet().size()
 		&& ImGui::Button("Actor Create"))
@@ -386,7 +385,7 @@ void GameEngineFBXWindow::OnGUI()
 
 			Renderer->SetFBXMesh(SelectMesh->GetName(), "Color");
 
-			for (size_t i = 0; i < Renderer->GetRenderSetCount(); i++)
+			for (int i = 0; i < Renderer->GetRenderSetCount(); i++)
 			{
 				Renderer->GetRenderSet(i).ShaderHelper->SettingConstantBufferSet("ResultColor", float4::RED);
 			}
